Remove old pools table (#2876)
* delete old pools table

* delete more

* move historical data store

* remove unused cells/cards

* remove unused localization keys

* lint

* update turbo
