+++ package.json
@@ -56,15 +56,15 @@
     "eslint-plugin-unicorn": "^48.0.1",
     "eslint-plugin-unused-imports": "^3.0.0",
     "jest": "^27.4.5",
+    "jest-watch-typeahead": "^2.2.2",
     "lerna": "^4.0.0",
     "lint-staged": "^12.1.3",
+    "msw": "^1.3.2",
     "pre-commit": "^1.2.2",
     "prettier": "^2.8.8",
     "ts-jest": "^27.1.2",
-    "turbo": "^1.10.16",
+    "turbo": "^1.12.4",
     "typescript": "4.7.4",
-    "jest-watch-typeahead": "^2.2.2",
-    "msw": "^1.3.2",
     "whatwg-fetch": "^3.6.19"
   },
   "dependencies": {},
