+++ packages/web/components/cards/index.ts
@@ -1,4 +1,3 @@
-export * from "./asset-card";
 export * from "./asset-source";
 export * from "./bond-card";
 export * from "./hero-card";
