+++ packages/web/components/complex/index.ts
@@ -1,2 +1,1 @@
-export * from "./paginated-table";
 export * from "./pool";
