+++ packages/web/components/table/cells/index.ts
@@ -1,9 +1,7 @@
 export * from "./asset-name";
 export * from "./balance";
-export * from "./metric-loader-cell";
 export * from "./pool-composition";
 export * from "./pool-quick-actions";
-export * from "./price-cell";
 export * from "./transfer-button";
 export * from "./types";
 export * from "./validator-info";
