+++ packages/web/localizations/zh-hk.json
@@ -282,13 +282,8 @@
       "transakSubtitle": "支援OSMO",
       "onrampMoneySubtitle": "支援OSMO、USDC"
     },
-    "incentive": {
-      "mobileTitle": "獎勵",
-      "title": "獎勵類型"
-    },
     "MAX": "全部",
     "pool": {
-      "mobileTitle": "流動性池",
       "title": "流動性池類型"
     },
     "searchTokens": "搜尋代幣",
@@ -312,9 +307,6 @@
       "weighted": "加權",
       "concentrated": "流動性集中",
       "transmuter": "轉換器",
-      "internal": "內部激勵",
-      "external": "外部激勵",
-      "superfluid": "超流體",
       "noIncentives": "無激勵"
     }
   },
@@ -489,7 +481,6 @@
   },
   "pools": {
     "allPools": {
-      "search": "搜尋流動性池",
       "sort": {
         "APRIncentivized": "年利率(APR)",
         "fees": "交易費 (7日)",
