--- packages/web/stores/derived-data/concentrated-liquidity/index.ts
@@ -1,1 +1,0 @@
-export * from "./historical-and-liquidity-data";
