--- packages/web/stores/derived-data/index.ts
@@ -1,3 +1,0 @@
-export * from "./concentrated-liquidity";
-export * from "./pools";
-export * from "./store";
