--- packages/web/stores/derived-data/pools/index.ts
@@ -1,2 +1,0 @@
-export * from "./metrics";
-export * from "./verified";
