+++ packages/web/stores/root.ts
@@ -8,6 +8,7 @@ import {
   ChainStore,
   CosmosAccount,
   CosmwasmAccount,
+  DerivedDataStore,
   IBCTransferHistoryStore,
   LPCurrencyRegistrar,
   NonIbcBridgeHistoryStore,
@@ -38,7 +39,6 @@ import { AxelarTransferStatusSource } from "~/integrations/bridges/axelar/axelar
 import { SkipTransferStatusSource } from "~/integrations/bridges/skip/skip-transfer-status-source";
 import { SquidTransferStatusSource } from "~/integrations/bridges/squid/squid-transfer-status-source";
 import { ObservableAssets } from "~/stores/assets";
-import { DerivedDataStore } from "~/stores/derived-data";
 import { makeIndexedKVStore, makeLocalStorageKVStore } from "~/stores/kv-store";
 import { NavBarStore } from "~/stores/nav-bar";
 import { ProfileStore } from "~/stores/profile";
@@ -235,9 +235,7 @@ export class RootStore {
       this.queriesExternalStore,
       this.accountStore,
       this.priceStore,
-      this.chainStore,
-      this.assetsStore,
-      this.userSettings
+      this.chainStore
     );
 
     this.ibcTransferHistoryStore = new IBCTransferHistoryStore(
